Refactor resource resolver; move it to WS and HTTP transport adapters
