+++ dh/transportadapter/transportadapter.go
@@ -12,4 +12,5 @@ func New(tsp transport.Transporter) TransportAdapter {
 
 type TransportAdapter interface {
 	HandleResponseError(rawRes []byte) error
+	ResolveResource(resName string, data map[string]interface{}) (resource, method string)
 }
